# gin-front-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "博客用户后台API接口文档",
        "title": "博客用户后台",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/login": {
            "post": {
                "description": "用户登录",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户相关接口"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.LoginDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/result.Result"
                        }
                    }
                }
            }
        },
...
        "/api/register": {
            "post": {
                "description": "用户注册",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户相关接口"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.RegisterDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/result.Result"
                        }
                    }
                }
            }
        },
...
        "/api/user/reset/password": {
            "post": {
                "description": "重置密码",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户相关接口"
                ],
                "summary": "重置密码",
                "parameters": [
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ResetPasswordDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/result.Result"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.LoginDto": {
            "type": "object",
            "required": [
                "checkCode",
                "email",
                "password"
            ],
            "properties": {
                "checkCode": {
                    "description": "验证码",
                    "type": "string",
                    "maxLength": 6,
                    "minLength": 4
                },
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "idKey": {
                    "description": "uuid",
                    "type": "string"
                },
                "password": {
                    "description": "密码",
                    "type": "string"
                }
            }
        },
...
        "model.RegisterDto": {
            "type": "object",
            "required": [
                "email",
                "emailCode",
                "password",
                "resetPassword",
                "username"
            ],
            "properties": {
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "emailCode": {
                    "description": "邮箱验证码",
                    "type": "string"
                },
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "resetPassword": {
                    "description": "重复密码",
                    "type": "string"
                },
                "username": {
                    "description": "用户账号",
                    "type": "string"
                }
            }
        },
        "model.ResetPasswordDto": {
            "type": "object",
            "required": [
                "email",
                "emailCode",
                "password",
                "resetPassword"
            ],
            "properties": {
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "emailCode": {
                    "description": "邮箱验证码",
                    "type": "string"
                },
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "resetPassword": {
                    "description": "重复密码",
                    "type": "string"
                }
            }
        },
...
        "result.Result": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "返回的数据"
                },
                "message": {
                    "description": "提示信息",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}